fix(tbc): address PR #1024 review comments
- Add comment documenting that inscription IDs use little-endian
  index per the ord protocol (AL-CT #5)
- Move ordinalindex_design.go to ordinalindex_design.md (AL-CT #11)
- Simplify OrdinalInscribedSatBounds using util.BytesPrefix and
  First()/Last() instead of manual seek logic (AL-CT #1)
- Move TestDbUpgradeV6 to level_test.go with other upgrade tests
  (AL-CT #14)
