+++ database/tbcd/level/level_test.go
@@ -1647,3 +1647,89 @@ func TestDbUpgradeV5Errors(t *testing.T) {
 		})
 	}
 }
+
+func TestDbUpgradeV6(t *testing.T) {
+	ctx, cancel := context.WithCancel(t.Context())
+	defer cancel()
+
+	home := t.TempDir()
+	cfg, err := NewConfig("upgradetest", home, "0mb", "0mb")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Phase 1: open with upgrade skipping, stamp version to 5, close.
+	cfg.SetUpgradeOpen(true)
+	db, err := New(ctx, cfg)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Seed a survivor row in MetadataDB to verify v6 doesn't wipe
+	// existing data.
+	survivorKey := []byte("v6-test-survivor")
+	survivorVal := []byte("must-survive-upgrade")
+	if err := db.MetadataPut(ctx, survivorKey, survivorVal); err != nil {
+		t.Fatal(err)
+	}
+
+	// Stamp version to 5.
+	v5 := make([]byte, 8)
+	binary.BigEndian.PutUint64(v5, 5)
+	if err := db.MetadataPut(ctx, versionKey, v5); err != nil {
+		t.Fatal(err)
+	}
+	if err := db.Close(); err != nil {
+		t.Fatal(err)
+	}
+
+	// Phase 2: reopen — the v5→v6 upgrade runs automatically.
+	cfg.SetUpgradeOpen(false)
+	db2, err := New(ctx, cfg)
+	if err != nil {
+		t.Fatalf("reopen (runs v6): %v", err)
+	}
+	defer func() {
+		if err := db2.Close(); err != nil {
+			t.Fatal(err)
+		}
+	}()
+
+	// Assertion 1: version is now 6.
+	ver, err := db2.Version(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if ver != 6 {
+		t.Fatalf("version: got %d, want 6", ver)
+	}
+
+	// Assertion 2: survivor data is intact.
+	got, err := db2.MetadataGet(ctx, survivorKey)
+	if err != nil {
+		t.Fatalf("survivor key lost: %v", err)
+	}
+	if string(got) != string(survivorVal) {
+		t.Fatalf("survivor value: got %q, want %q", got, survivorVal)
+	}
+
+	// Assertion 3: ordinal DB is functional — can write and read back.
+	txid := chainhash.Hash{0x01, 0x02}
+	outpoint := tbcd.NewOutpoint(txid, 0)
+	var rKey tbcd.OrdinalKey
+	rKey[0] = 'r'
+	copy(rKey[1:], outpoint[:])
+	cache := map[tbcd.OrdinalKey]tbcd.OrdinalValue{
+		rKey: tbcd.OrdinalValue([]byte{0xaa, 0xbb}),
+	}
+	if err := db2.BlockOrdinalUpdate(ctx, 1, cache, chainhash.Hash{}); err != nil {
+		t.Fatalf("BlockOrdinalUpdate after upgrade: %v", err)
+	}
+	got2, err := db2.OrdinalSatRangesByOutpoint(ctx, outpoint)
+	if err != nil {
+		t.Fatalf("OrdinalSatRangesByOutpoint after upgrade: %v", err)
+	}
+	if len(got2) != 2 || got2[0] != 0xaa || got2[1] != 0xbb {
+		t.Fatalf("ordinal round-trip after upgrade: got %x", got2)
+	}
+}
